app/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbot"
version = "0.1.0"
description = "A chatbot with LangGraph"
readme = "README.md"
authors = [
    {name = "Your Name"}
]
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-tavily>=0.1.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sse-starlette>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "langgraph-cli[inmem]",
]

[tool.setuptools]
packages = ["chatbot"] 